WrapperBackend in wrap_compiler_fn(); move testing.same() to utils.py
